Fix command parsing to use the bot_command entity
Telegram sends a `bot_command` entity for commands highlighted by the
client, such as `/command` or `/command@botname`. This commit makes the
library check whether there's a `bot_command` entity at offset 0 instead
of simply checking if the message starts with `/` to ensure that
formatted text beginning with a slash isn't erroneously interpreted as
a command.